+++ test/config/jsfun.js
@@ -1,0 +1,10 @@
+module.exports = function(grunt, options) {
+  return {
+    jsFunFile: {
+      options: {
+        filename: 'jsfun.js',
+        test: options.test
+      }
+    }
+  };
+};